refactor: isolate network I/O and fix async runtime conflicts - genesis-ide: Isolated telemetry and geometry fetchers into dedicated OS threads with independent Tokio runtimes. This prevents "no reactor running" panics within Bevy's ComputeTaskPool. - genesis-ide: Patched handle_picking and handle_box_select systems to use Option<Res<LoadedGeometry>>, preventing ECS access panics during cold boot/loading states. - genesis-node: Fixed EADDRINUSE fatal panic in IoMultiplexer. Port 8081 binding now fails gracefully with a log entry instead of crashing the daemon. - genesis-node: Removed nested tokio::Runtime from ZoneRuntime to resolve "cannot drop a runtime in async context" errors during boot/shutdown. - genesis-baker: (prev) Fixed type mismatches in manifest variant mapping for i32/u8 compatibility.

+++ genesis-baker/src/main.rs
@@ -348,14 +348,14 @@ fn compile(sim_path: &Path, bp_path: &Path, an_path: &Path, io_path: &Path, out_
                 rest_potential: v.rest_potential,
                 leak_rate: v.leak_rate,
                 homeostasis_penalty: v.homeostasis_penalty,
-                homeostasis_decay: v.homeostasis_decay,
-                gsop_potentiation: v.gsop_potentiation,
-                gsop_depression: v.gsop_depression,
+                homeostasis_decay: v.homeostasis_decay as i32,
+                gsop_potentiation: v.gsop_potentiation as i32,
+                gsop_depression: v.gsop_depression as i32,
                 refractory_period: v.refractory_period,
                 synapse_refractory_period: v.synapse_refractory_period,
                 slot_decay_ltm: v.slot_decay_ltm as u8,
                 slot_decay_wm: v.slot_decay_wm as u8,
-                signal_propagation_length: v.signal_propagation_length as u16,
+                signal_propagation_length: v.signal_propagation_length as u8,
             }
         }).collect(),
     };
